test: clean up duplicate db fixtures and minor import fixes
- Move `test_engine` and `test_db_path` fixtures to `tests/conftest.py`
- Remove redundant local definitions from `tests/core/db/test_connection.py`
- Fix casing issue in authorization header check in `test_http_auth.py`
- Minor import ordering and comment cleanup in metadata service
